Move server2 back into server1
